requirements.txt
flask
flask-cors
python-dotenv
wikipedia
google-generativeai
openai
